Split main.ts into focused modules (847→642 lines)
Extract three focused modules from main.ts:
- tauri-bootstrap.ts (60 lines): waitForTauri() function
- engine-handlers.ts (178 lines): Engine start/reset/stop logic
- terminal-display.ts (166 lines): xterm.js terminal initialization

This reduces main.ts by ~24% (205 lines) while improving:
- Maintainability: Each module has clear, single responsibility
- Testing: Can unit test initialization modules independently
- Readability: Easier to find and modify specific functionality

Closes #797
